Allow non-core account links to appear in menu
Currently, any community or local modules which try to append links to
this menu are being hidden, this patch allows them to show below the
configured menu. The starting position for the custom links is 150. All
of the core links start with their position below 150 and can be
configured to before or after any other link. They cannot as yet be
turned off though.